update env creation
