+++ pyproject.toml
@@ -13,7 +13,6 @@ classifiers=[
     ]
 keywords=['systems', 'biology', 'modeling', 'simulation']
 
-
 [tool.poetry.dependencies]
 python = "^3.10"
