+++ worker/service/main.py
@@ -1,13 +1,49 @@
 import os
 import asyncio
 import logging
+from typing import List, Tuple
 
+from process_bigraph import ProcessTypes
 from dotenv import load_dotenv
 
-from service import APP_PROCESS_REGISTRY
 from service.shared_worker import MongoDbConnector
 from service.log_config import setup_logging
 from service.job import Supervisor
+
+
+def register_module(
+        items_to_register: List[Tuple[str, str]],
+        core: ProcessTypes,
+        verbose=False
+) -> None:
+    for process_name, path in items_to_register:
+        module_name, class_name = path.rsplit('.', 1)
+        try:
+            import_statement = f'worker.service.bigraph.{module_name}'
+
+            module = __import__(
+                 import_statement, fromlist=[class_name])
+
+            # Get the class from the module
+            bigraph_class = getattr(module, class_name)
+
+            # Register the process
+            core.process_registry.register(process_name, bigraph_class)
+            print(f'Registered {process_name}') if verbose else None
+        except Exception as e:
+            print(f"Cannot register {class_name}. Error:\n**\n{e}\n**") if verbose else None
+            continue
+
+
+APP_PROCESS_REGISTRY = ProcessTypes()
+IMPLEMENTATIONS = [
+    ('output-generator', 'steps.OutputGenerator'),
+    ('time-course-output-generator', 'steps.TimeCourseOutputGenerator'),
+    # ('smoldyn_step', 'steps.SmoldynStep'),
+    ('simularium_smoldyn_step', 'steps.SimulariumSmoldynStep'),
+    ('mongo-emitter', 'steps.MongoDatabaseEmitter')
+]
+register_module(IMPLEMENTATIONS, APP_PROCESS_REGISTRY, verbose=True)
 
 
 load_dotenv('../../assets/dev/config/.env_dev')
