+++ worker/service/shared_worker.py
@@ -17,6 +17,7 @@ from pymongo.database import Database
 
 # -- globally-shared content-- #
 
+
 load_dotenv('../../assets/dev/config/.env_dev')
 
 DB_TYPE = "mongo"  # ie: postgres, etc
